InterfaceChange_Ver1
Thêm bootstrap và menu

+++ src/main/java/com/example/asmjava5/DAO/ProductServiceImpl.java
@@ -25,7 +25,7 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public Product findById(Long id) {
+    public Product findById(Integer id) {
         return productRepository.findById(id).orElse(null);
     }
 
@@ -35,7 +35,7 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public void delete(Long id) {
+    public void delete(Integer id) {
         productRepository.deleteById(id);
     }
